chore: move screens to an appropriate package
