+++ app/src/main/java/com/pereyrarg11/mobile/core/presentation/screen/FeedbackScreen.kt
@@ -1,4 +1,4 @@
-package com.pereyrarg11.mobile.core.presentation.components
+package com.pereyrarg11.mobile.core.presentation.screen
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
