+++ app/src/main/java/com/pereyrarg11/mobile/core/presentation/screen/LoadingScreen.kt
@@ -1,0 +1,42 @@
+package com.pereyrarg11.mobile.core.presentation.screen
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.tooling.preview.Preview
+import com.pereyrarg11.mobile.R
+import com.pereyrarg11.mobile.core.presentation.theme.WhiteLabelTheme
+import com.pereyrarg11.mobile.core.presentation.util.UiText
+
+@Composable
+fun LoadingScreen(
+    modifier: Modifier = Modifier,
+    message: UiText = UiText.StringResource(R.string.message_loading),
+) {
+    Box(
+        modifier = modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center,
+    ) {
+        Column(horizontalAlignment = Alignment.CenterHorizontally) {
+            CircularProgressIndicator()
+            Spacer(modifier = Modifier.height(dimensionResource(id = R.dimen.size_lg)))
+            Text(text = message.asString())
+        }
+    }
+}
+
+@Preview
+@Composable
+fun LoadingScreenPreview() {
+    WhiteLabelTheme {
+        LoadingScreen()
+    }
+}
